Refactor downloader to use yt-dlp-wrap
This commit improves the project in two main ways:

1. The core downloading logic in `src/services/downloader.ts` has been refactored to use the `yt-dlp-wrap` library instead of directly spawning a `yt-dlp` child process. This makes the project more robust and easier to maintain by abstracting away the process management. It also adds `yt-dlp-wrap` as a dependency in `package.json`, so `yt-dlp` is managed as part of the project and doesn't need to be installed globally by the user.

2. A typo in the filename `dowloader.ts` has been corrected to `downloader.ts`, and the corresponding import in `src/index.ts` has been updated.
